Move cache functions to own module + print relative paths to log + fix rerun in same session
