+++ src/protein_detective/workflow.py
@@ -1,29 +1,27 @@
 """Workflow steps"""
 
-import dataclasses
-import hashlib
-import json
 import logging
-from dataclasses import asdict, dataclass
+from dataclasses import dataclass
 from pathlib import Path
 from typing import Literal, cast
 
 from dask.distributed import Client, progress
 from distributed.deploy.cluster import Cluster
-from platformdirs import user_cache_dir
 
-from protein_detective.alphafold import AlphaFoldEntry, DownloadableFormat
+from protein_detective.alphafold import DownloadableFormat
 from protein_detective.alphafold import fetch_many as af_fetch
 from protein_detective.alphafold import relative_to as af_relative_to
 from protein_detective.alphafold.density import DensityFilterQuery, filter_on_density
+from protein_detective.cache import (
+    try_linking_files_from_another_session,
+    try_reusing_search_results_from_another_session,
+)
 from protein_detective.db import (
     connect,
-    copy_search_results,
     load_alphafold_ids,
     load_alphafolds,
     load_pdb_ids,
     load_pdbs,
-    load_uniprot_queries,
     save_alphafolds,
     save_alphafolds_files,
     save_density_filtered,
@@ -46,17 +44,6 @@ from protein_detective.uniprot import Query, search4af, search4pdb, search4unipr
 
 logger = logging.getLogger(__name__)
 
-cache_dir = Path(user_cache_dir("protein-detective", ensure_exists=True))
-cache_uniprot_root = cache_dir / "uniprot"
-cache_uniprot_root.mkdir(parents=True, exist_ok=True)
-
-
-def _generate_query_hash(query: Query, limit: int) -> str:
-    query_dict = asdict(query)
-    query_dict["limit"] = limit
-    body = json.dumps(query_dict, sort_keys=True, separators=(",", ":"))
-    return hashlib.sha256(body.encode("utf-8")).hexdigest()
-
 
 def search_structures_in_uniprot(query: Query, session_dir: Path, limit: int = 10_000) -> tuple[int, int, int, int]:
     """Searches for protein structures in UniProt database.
@@ -76,7 +63,7 @@ def search_structures_in_uniprot(query: Query, session_dir: Path, limit: int = 1
     """
     session_dir.mkdir(parents=True, exist_ok=True)
 
-    _try_reused_cached_search_results(query, session_dir, limit)
+    try_reusing_search_results_from_another_session(query, session_dir, limit)
 
     # Reuse existing search results in the current session database
     with connect(session_dir) as con:
@@ -102,120 +89,10 @@ def search_structures_in_uniprot(query: Query, session_dir: Path, limit: int = 1
     return len(uniprot_accessions), nr_pdbs, nr_prot2pdb, nr_afs
 
 
-def _try_reused_cached_search_results(query, session_dir, limit):
-    qhash = _generate_query_hash(query, limit)
-    cache_path = cache_uniprot_root / qhash
-    if cache_path.exists() and cache_path.readlink().exists():
-        cached_session_dir = cache_path.readlink()
-        if session_dir.absolute() == cached_session_dir.absolute():
-            logger.warning("Running same query in this session dir again.")
-        else:
-            with connect(session_dir) as con:
-                logger.warning(f"Query seen before, copying search results from {cached_session_dir} session.")
-                copy_search_results(cached_session_dir, con)
-    else:
-        cache_path.unlink(missing_ok=True)
-        cache_path.symlink_to(session_dir.absolute())
-
-
 WhatRetrieve = Literal["pdbe", "alphafold"]
 """Types of what to retrieve."""
 what_retrieve_choices: set[WhatRetrieve] = {"pdbe", "alphafold"}
 """Set of what can be retrieved."""
-
-
-def _try_linking_files_from_another_session(
-    session_dir: Path, what: set[WhatRetrieve], what_af_formats: set[DownloadableFormat], download_dir: Path
-) -> tuple[dict[str, Path], list[AlphaFoldEntry]]:
-    # Check if there is an another session with same query and downloaded files
-    with connect(session_dir, read_only=True) as con:
-        uniprot_queries = load_uniprot_queries(con)
-    if not uniprot_queries:
-        msg = "No UniProt queries results found in the session database. Please run a search first."
-        raise ValueError(msg)
-    query, limit = uniprot_queries[0]
-    if len(uniprot_queries) > 1:
-        logger.warning(
-            "More than one UniProt query found in the session database. "
-            "Only using the first one to see if it was retrieved in another session"
-        )
-
-    qhash = _generate_query_hash(query, limit)
-    cached_session_dir = cache_uniprot_root / qhash
-    linked_files = {}
-    current_afs = []
-    if (
-        not cached_session_dir.exists()
-        or not cached_session_dir.readlink().exists()
-        or cached_session_dir == session_dir
-    ):
-        logger.info("No cached session found for query, downloading files instead of linking to existing files.")
-        return linked_files, current_afs
-    logger.info(
-        'Found cached session "%s" with same query, linking files instead of re-downloading', cached_session_dir
-    )
-    cached_session_dir = cached_session_dir.readlink().relative_to(session_dir.absolute(), walk_up=True)
-    if "pdbe" in what:
-        linked_files = _symlink_cached_pdbe_files(session_dir, download_dir, cached_session_dir)
-    if "alphafold" in what:
-        current_afs = _symlink_cached_alphafold_files(session_dir, what_af_formats, download_dir, cached_session_dir)
-    return linked_files, current_afs
-
-
-def _camel_to_snake_case(name: str) -> str:
-    """Convert a camelCase string to snake_case."""
-    return "".join(["_" + c.lower() if c.isupper() else c for c in name]).lstrip("_")
-
-
-def _symlink_cached_alphafold_files(
-    session_dir: Path, what_af_formats: set[DownloadableFormat], download_dir: Path, cached_session_dir: Path
-) -> list[AlphaFoldEntry]:
-    with connect(cached_session_dir, read_only=True) as cache_con:
-        cached_afs = load_alphafolds(cache_con)
-    if cached_afs:
-        logger.warning(
-            'AlphaFold files already downloaded in session "%s". Symlinking instead of downloading.', cached_session_dir
-        )
-    current_afs: list[AlphaFoldEntry] = []
-    for af in cached_afs:
-        af_files = {}
-        for af_format in what_af_formats:
-            # DownloadableFormat uses camelcAse, while AlphaFoldEntry uses snake_case, so convert
-            entry_key = _camel_to_snake_case(af_format) + "_file"
-            cached_file = getattr(af, entry_key)
-            if cached_file is None:
-                continue
-            linked_file = download_dir / cached_file.name
-            if not linked_file.exists():
-                rel_cached_file = cached_file.resolve().relative_to(download_dir.absolute(), walk_up=True)
-                linked_file.symlink_to(rel_cached_file)
-                af_files[entry_key] = linked_file.relative_to(session_dir)
-        current_af = dataclasses.replace(af, **af_files)
-        current_afs.append(current_af)
-    with connect(session_dir) as scon:
-        save_alphafolds_files(current_afs, scon)
-    return current_afs
-
-
-def _symlink_cached_pdbe_files(session_dir: Path, download_dir: Path, cached_session_dir: Path) -> dict[str, Path]:
-    with connect(cached_session_dir, read_only=True) as cache_con:
-        cached_pdbes = load_pdbs(cache_con)
-    if cached_pdbes:
-        logger.warning(
-            'PDBe files already downloaded in session "%s". Symlinking instead of downloading.', cached_session_dir
-        )
-    linked_files: dict[str, Path] = {}
-    for pdb in cached_pdbes:
-        if pdb.mmcif_file:
-            current_pdb = download_dir / pdb.mmcif_file.name
-            if current_pdb.exists():
-                continue
-            rel_cached_file = pdb.mmcif_file.resolve().relative_to(download_dir.absolute(), walk_up=True)
-            current_pdb.symlink_to(rel_cached_file)
-            linked_files[pdb.id] = current_pdb.relative_to(session_dir)
-    with connect(session_dir) as scon:
-        save_pdb_files(linked_files, scon)
-    return linked_files
 
 
 def retrieve_structures(
@@ -247,9 +124,7 @@ def retrieve_structures(
     if what_af_formats is None:
         what_af_formats = {"pdb"}
 
-    linked_files, current_afs = _try_linking_files_from_another_session(
-        session_dir, what, what_af_formats, download_dir
-    )
+    linked_files, current_afs = try_linking_files_from_another_session(session_dir, what, what_af_formats, download_dir)
     if linked_files or current_afs:
         return download_dir, len(linked_files), len(current_afs)
